Moving main to the user interface class.
