+++ inventory/src/UserInterface.java
@@ -1,5 +1,6 @@
+import java.io.File;
+import java.io.FileNotFoundException;
 import java.util.ArrayList;
-import java.util.Locale;
 import java.util.Scanner;
 
 public class UserInterface {
@@ -8,6 +9,24 @@ public class UserInterface {
 
     public void changeModule() {
 
+    }
+
+    public static void main(String[] args) throws FileNotFoundException {
+        Product p = new Product();
+        p.setName("Turner's Fantastic Stir Fry and Cherry Pie");
+        p.setPrice(19.99);
+        p.setDescription("This Stir Fry and Cherry Pie evokes memories of the salty breeze off the ocean, and a warm summer day");
+        //System.out.println(p.getName() + " \nPrice: " + p.getPrice() + " \nDescription: " + p.getDescription());
+
+        File data = new File("data.csv");
+
+        Inventory inventory = Inventory.getInstance();
+        inventory.addProduct(p);
+
+        inventory.importProducts(data);
+        DisplayInventory.display(inventory);
+
+        UserInterface.menu(inventory);
     }
 
     public static void menu(Inventory inv) {
